Move, un-env-ify some task subroutines
Re #21
